arrange functions into modules with io dir
